requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
aiosqlite
asyncpg
python-jose
passlib[bcrypt]
pydantic-settings

greenlet
pydantic[email]